readme: update setup-tarantool action URL
It was moved from https://github.com/rosik/setup-tarantool into
tarantool organization on GitHub.

The old URL redirects to the new one, but there is no guarantee that
it'll be so in a future. I guess if Yaroslav will fork the
tarantool/setup-tarantool repository, rosik/setup-tarantool will point
to the fork. It is more safe to use the new URL.

Reported by Vladislav Grubov (@ochaton).