Bump Flask from 1.1.1 to 2.2.2
This patch is intended to fix the following issue:

    E   ImportError: cannot import name 'escape' from 'jinja2'

Jinja is a dependency of Flask and Flask 1.x.x uses the escape module
from Jinja, however recently support for the escape module was dropped
in newer versions of Jinja.

To fix this issue, simply update to the newer version of Flask 2.x.x
where Flask no longer uses the escape module from Jinja.